+++ packages/react-native/ReactAndroid/proguard-rules.pro
@@ -48,7 +48,7 @@
 
 -dontwarn com.facebook.react.**
 -keep,includedescriptorclasses class com.facebook.react.bridge.** { *; }
--keep,includedescriptorclasses class com.facebook.react.turbomodule.core.** { *; }
+-keep,includedescriptorclasses class com.facebook.react.internal.turbomodule.core.** { *; }
 
 # hermes
 -keep class com.facebook.jni.** { *; }
