+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/module/processing/ReactModuleSpecProcessor.java
@@ -161,13 +161,13 @@ public class ReactModuleSpecProcessor extends ProcessorBase {
       builder.addStatement("$T map = new $T()", MAP_TYPE, INSTANTIATED_MAP_TYPE);
 
       String turboModuleInterfaceCanonicalName =
-          "com.facebook.react.turbomodule.core.interfaces.TurboModule";
+          "com.facebook.react.internal.turbomodule.core.interfaces.TurboModule";
       TypeMirror turboModuleInterface =
           mElements.getTypeElement(turboModuleInterfaceCanonicalName).asType();
 
       if (turboModuleInterface == null) {
         throw new RuntimeException(
-            "com.facebook.react.turbomodule.core.interfaces.TurboModule interface not found.");
+            "com.facebook.react.internal.turbomodule.core.interfaces.TurboModule interface not found.");
       }
 
       for (String nativeModule : nativeModules) {
