+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/modules/fresco/FrescoModule.java
@@ -19,12 +19,12 @@ import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.common.ReactConstants;
+import com.facebook.react.internal.turbomodule.core.interfaces.TurboModule;
 import com.facebook.react.module.annotations.ReactModule;
 import com.facebook.react.modules.common.ModuleDataCleaner;
 import com.facebook.react.modules.network.CookieJarContainer;
 import com.facebook.react.modules.network.ForwardingCookieHandler;
 import com.facebook.react.modules.network.OkHttpClientProvider;
-import com.facebook.react.turbomodule.core.interfaces.TurboModule;
 import java.util.HashSet;
 import okhttp3.JavaNetCookieJar;
 import okhttp3.OkHttpClient;
