+++ packages/react-native/ReactAndroid/src/test/java/com/facebook/react/bridge/BaseJavaModuleTest.kt
@@ -7,7 +7,7 @@
 
 package com.facebook.react.bridge
 
-import com.facebook.react.turbomodule.core.interfaces.TurboModule
+import com.facebook.react.internal.turbomodule.core.interfaces.TurboModule
 import com.facebook.testutils.shadows.ShadowSoLoader
 import org.junit.Before
 import org.junit.Test
